refactor(orulo): extract building payloads processor
